fix(dashboard): keep the session view still when a run stops
The page swapped RunLive for RunReplay the instant a run's status
flipped, so everything remounted at once: the action bar blanked while
its git read went out again, the output was replaced by "Loading
session…" while the archived log was fetched, the run overview
disappeared, and the composer was rebuilt. A session ending is the
moment you are most likely to be reading it.

Live and finished are now one view (RunView) with a `live` flag. The
bar, feed and composer stay mounted; the archived log is read behind the
events already on screen. The two composers become one (RunComposer):
the input never moves, and only what a send does changes — a message
into a running session, a --resume continuation of a finished one, or a
new session when there is no id to resume, which used to remove the
composer entirely and leave a dead end.

The action bar is one row at any width: the branch truncates, the size
and summary drop out as the pane narrows, and the buttons never wrap
under them.

Closes #1026
