+++ packages/framework-dashboard/components/RunFeed.tsx
@@ -4,7 +4,7 @@ import { EventList } from './EventList.js'
 import { RunOverview } from './RunOverview.js'
 
 // One run's feed: the run overview plus the live/replayed event log, or a waiting placeholder
-// before anything has streamed. Shared by the run's own view (RunLive, which shows the session
+// before anything has streamed. Shared by the run's own view (RunView, which shows the session
 // link in its action bar instead — `showSessionLink={false}`) and the read-only relay watch view
 // (RelayView, which keeps it since it has no action bar). `lost` is the live channel's health
 // (#948): while the stream is down the feed is behind reality, and saying so beats letting
@@ -13,10 +13,18 @@ export function RunFeed({
   events,
   showSessionLink = true,
   lost = false,
+  stick = true,
+  openAt,
+  emptyLabel = 'Waiting for the session to start…',
 }: {
   events: FrameworkEvent[]
   showSessionLink?: boolean
   lost?: boolean
+  /** A finished log is static (#1026): it does not follow new output, and opens at its end. */
+  stick?: boolean
+  openAt?: 'start' | 'end'
+  /** What an empty feed says: a live run is waiting, a finished one has nothing to replay. */
+  emptyLabel?: string
 }) {
   const lostBanner = lost && (
     <div role="status" className="flex items-center gap-2 border-b border-border bg-warning/10 px-4 py-2 text-xs text-warning">
@@ -28,7 +36,7 @@ export function RunFeed({
     return (
       <>
         {lostBanner}
-        <div className="grid flex-1 place-items-center text-sm text-muted-foreground">Waiting for the session to start…</div>
+        <div className="grid flex-1 place-items-center text-sm text-muted-foreground">{emptyLabel}</div>
       </>
     )
   }
@@ -36,7 +44,7 @@ export function RunFeed({
     <>
       {lostBanner}
       <RunOverview events={events} showSessionLink={showSessionLink} />
-      <EventList events={events} />
+      <EventList events={events} stick={stick} {...(openAt ? { openAt } : {})} />
     </>
   )
 }
